+++ CHANGELOG.md
@@ -28,6 +28,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Only try to save Azure VMSS IDs in Custom Resources if VMSS exists.
 - Fix firewall rules to allow traffic between nodes in all node pools.
 
+### Changed
+
+- Use `AzureCluster.Spec.NetworkSpec.Vnet.CIDRBlocks` field instead of deprecated `AzureCluster.Spec.NetworkSpec.Vnet.CidrBlock`.
+- Use `Subnet.CIDRBlocks` field instead of deprecated `Subnet.CidrBlock`.
+
 ## [5.0.0-alpha4] - 2020-10-27
 
 ### Changed